move to_json logic all out of ruleset
