+++ tests/bv4.rs
@@ -49,10 +49,11 @@ pub mod test {
         let rules = bv4_rules();
         let duration = start.elapsed();
 
-        rules.write_json_rules("bv4.json");
+        logger::write_json_rules(&rules, "bv4.json");
         let baseline = Ruleset::<_>::from_file("baseline/bv4.rules");
-        rules.write_output(
-            baseline,
+        logger::write_output(
+            &rules,
+            &baseline,
             "bv4",
             "oopsla",
             "baseline.json",
