+++ tests/rational.rs
@@ -578,9 +578,10 @@ pub mod test {
         let rules = replicate_ruler1_recipe();
         let duration = start.elapsed();
 
-        rules.write_json_rules("rational_replicate.json");
-        rules.write_output(
-            ruler1.clone(),
+        logger::write_json_rules(&rules, "rational_replicate.json");
+        logger::write_output(
+            &rules,
+            &ruler1,
             "rational_replicate",
             "oopsla",
             "baseline.json",
@@ -590,8 +591,9 @@ pub mod test {
             },
             duration,
         );
-        rules.write_output(
-            herbie.clone(),
+        logger::write_output(
+            &rules,
+            &herbie,
             "rational_replicate",
             "herbie",
             "herbie.json",
@@ -606,9 +608,10 @@ pub mod test {
         let rules = best_enumo_recipe();
         let duration = start.elapsed();
 
-        rules.write_json_rules("rational_best.json");
-        rules.write_output(
-            ruler1,
+        logger::write_json_rules(&rules, "rational_best.json");
+        logger::write_output(
+            &rules,
+            &ruler1,
             "rational_best",
             "oopsla",
             "baseline.json",
@@ -618,8 +621,9 @@ pub mod test {
             },
             duration,
         );
-        rules.write_output(
-            herbie,
+        logger::write_output(
+            &rules,
+            &herbie,
             "rational_replicate",
             "herbie",
             "herbie.json",
